add moderator page-content request workflow + admin preview
- New admin Requests tab (approve/decline) and a moderator-facing
  Requests panel on the profile page; both link to a shared preview
  page that renders the real Home/About components against a
  request's proposed content instead of a mockup
- ImageWithFallback shows a placeholder instead of a broken image
  icon when a previewed image no longer exists (e.g. declined
  requests whose staged images were deleted)
- Split ProfilePage into components/profile/*, mirroring the
  existing components/admin/* pattern (~1700 lines down to ~350)
- New reusable FloatingBar/PageEditBar components for edit toolbars
  that hover over the page instead of reserving scroll space;
  ConfirmModal gained a danger variant to dedupe hand-rolled modals
- Admin dashboard's active tab now syncs to a ?tab= URL param
- Misc layout fixes (hero panel sizing, PageContainer vertical
  centering, mobile navbar height) and Tailwind v4 class cleanup
