+++ src/pages/ProfilePage.jsx
@@ -2,26 +2,15 @@ import React, { useEffect, useState, useMemo } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { isAuthenticated, fetchWithAuth } from "../auth/auth";
-import { optimizeCloudinaryUrl, toSlug, validatePhone } from "../util/util";
 import { Button, Spinner } from "../components/ui";
 import MyTeamRow from "../components/teams/MyTeamRow";
+import OrderRow from "../components/profile/OrderRow";
+import EnrollmentRow from "../components/profile/EnrollmentRow";
+import VenueBookingRow from "../components/profile/VenueBookingRow";
+import EventSubscriptionRow from "../components/profile/EventSubscriptionRow";
+import MyContentRequestsPanel from "../components/profile/MyContentRequestsPanel";
 import { API } from "../api/apiClient";
 import { queryKeys } from "../api/queryKeys";
-
-function formatDate(dateStr) {
-  if (!dateStr) return "—";
-  return new Date(dateStr).toLocaleDateString("en-GB", {
-    day: "numeric",
-    month: "short",
-    year: "numeric",
-  });
-}
-
-function formatCurrency(amount) {
-  return "£" + Number(amount ?? 0).toFixed(2);
-}
-
-const INTERVAL_ADJ = { month: "Monthly", year: "Yearly", week: "Weekly" };
 
 const TABS = ["Tickets", "Teams", "Enrollments", "Venues"];
 
@@ -70,7 +59,7 @@ export default function ProfilePage() {
 
   if (loading)
     return (
-      <div className="min-h-screen flex items-center justify-center bg-gradient-to-br from-base-200 via-white to-base-200">
+      <div className="min-h-screen flex items-center justify-center bg-linear-to-br from-base-200 via-white to-base-200">
         <div className="flex flex-col items-center gap-3">
           <Spinner />
           <p className="text-sm text-base-content/50">Loading your profile…</p>
@@ -80,7 +69,7 @@ export default function ProfilePage() {
 
   if (error)
     return (
-      <div className="min-h-screen flex items-center justify-center bg-gradient-to-br from-base-200 via-white to-base-200">
+      <div className="min-h-screen flex items-center justify-center bg-linear-to-br from-base-200 via-white to-base-200">
         <div className="bg-white rounded-2xl p-8 shadow-sm border border-red-100 text-center max-w-sm">
           <p className="text-red-500 font-medium mb-4">{error}</p>
           <Link to="/" className="text-sm text-base-content/70 hover:underline">
@@ -91,6 +80,7 @@ export default function ProfilePage() {
     );
 
   const { user, teams, enrollments = [], eventSubscriptions = [], venueBookings = [] } = data;
+  const visibleTabs = user.role === "moderator" ? [...TABS, "Requests"] : TABS;
   const initials = user.name
     ? user.name
         .split(" ")
@@ -101,10 +91,10 @@ export default function ProfilePage() {
     : "?";
 
   return (
-    <div className="min-h-screen bg-gradient-to-br from-base-200 via-white to-base-200">
+    <div className="min-h-screen bg-linear-to-br from-base-200 via-white to-base-200">
       <div className="max-w-4xl mx-auto px-4 pt-8">
         <div className="mb-6 flex items-center gap-5">
-          <div className="w-16 h-16 rounded-full bg-gradient-to-br from-primary to-primary/70 flex items-center justify-center text-white text-2xl font-bold shadow-md shadow-primary/20 flex-shrink-0">
+          <div className="w-16 h-16 rounded-full bg-linear-to-br from-primary to-primary/70 flex items-center justify-center text-white text-2xl font-bold shadow-md shadow-primary/20 shrink-0">
             {initials}
           </div>
 
@@ -152,7 +142,7 @@ export default function ProfilePage() {
         {/* ── Tabs ── */}
         <div className="border-b border-base-300 mb-6">
           <div className="flex gap-1">
-            {TABS.map((tab) => (
+            {visibleTabs.map((tab) => (
               <button
                 key={tab}
                 onClick={() => setActiveTab(tab)}
@@ -163,21 +153,23 @@ export default function ProfilePage() {
                 }`}
               >
                 {tab}
-                <span
-                  className={`ml-2 text-xs px-1.5 py-0.5 rounded-full ${
-                    activeTab === tab
-                      ? "bg-base-200 text-primary"
-                      : "bg-base-200 text-base-content/50"
-                  }`}
-                >
-                  {tab === "Tickets"
-                    ? Tickets.length
-                    : tab === "Teams"
-                      ? teams.length
-                      : tab === "Venues"
-                        ? venueBookings.length
-                        : enrollments.length + eventSubscriptions.length}
-                </span>
+                {tab !== "Requests" && (
+                  <span
+                    className={`ml-2 text-xs px-1.5 py-0.5 rounded-full ${
+                      activeTab === tab
+                        ? "bg-base-200 text-primary"
+                        : "bg-base-200 text-base-content/50"
+                    }`}
+                  >
+                    {tab === "Tickets"
+                      ? Tickets.length
+                      : tab === "Teams"
+                        ? teams.length
+                        : tab === "Venues"
+                          ? venueBookings.length
+                          : enrollments.length + eventSubscriptions.length}
+                  </span>
+                )}
               </button>
             ))}
           </div>
@@ -334,1285 +326,10 @@ export default function ProfilePage() {
                 ))}
               </div>
             ))}
+
+          {activeTab === "Requests" && <MyContentRequestsPanel />}
         </div>
       </div>
-    </div>
-  );
-}
-
-function OrderRow({ tickets }) {
-  const [expanded, setExpanded] = useState(false);
-  const firstTicket = tickets[0];
-  const event = firstTicket.eventId;
-  const qty = tickets.length;
-  const totalPaid = (event?.ticketPrice ?? 0) * qty;
-
-  return (
-    <div className="bg-white rounded-2xl border border-base-300 shadow-sm hover:shadow-md transition-all duration-200 overflow-hidden">
-      {/* Main row */}
-      <div className="flex">
-        {/* Date block */}
-        <div className="w-16 flex-shrink-0 bg-gradient-to-b from-primary to-primary/70 flex flex-col items-center justify-center text-white py-4">
-          <span className="text-xs font-semibold uppercase opacity-80">
-            {event?.date ? new Date(event.date).toLocaleString("en-GB", { month: "short" }) : "—"}
-          </span>
-          <span className="text-2xl font-bold leading-tight">
-            {event?.date ? new Date(event.date).getDate() : "—"}
-          </span>
-        </div>
-
-        {/* Event image */}
-        {event?.images?.[0] ? (
-          <img
-            src={optimizeCloudinaryUrl(event.images[0])}
-            alt={event.title}
-            className="w-28 object-cover flex-shrink-0"
-            width="112"
-            height="88"
-          />
-        ) : (
-          <div className="w-28 bg-gradient-to-br from-base-200 to-base-200 flex items-center justify-center flex-shrink-0">
-            <svg
-              className="w-8 h-8 text-base-content/30"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={1.5}
-                d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z"
-              />
-            </svg>
-          </div>
-        )}
-
-        {/* Info */}
-        <div className="flex-1 px-5 py-4 flex flex-col justify-center min-w-0">
-          <p className="font-semibold text-base-content truncate text-base">
-            {event?.title ?? "Unknown Event"}
-          </p>
-          <p className="text-sm text-base-content/50 mt-0.5">
-            {formatDate(event?.date)}
-            {event?.city && <span> · {event.city}</span>}
-          </p>
-          <p className="text-xs text-base-content/50 mt-1 font-mono">
-            {qty} ticket{qty !== 1 ? "s" : ""}
-          </p>
-        </div>
-
-        {/* Price + status */}
-        <div className="flex flex-col items-end justify-center px-5 gap-2 flex-shrink-0">
-          <span className="text-sm font-semibold text-base-content">
-            {formatCurrency(totalPaid)}
-          </span>
-          <span className="text-xs font-medium px-2.5 py-0.5 rounded-full bg-green-50 text-green-700 border border-green-200">
-            ✓ Paid
-          </span>
-        </div>
-      </div>
-
-      {/* Expand toggle for individual tickets */}
-      <div className="border-t border-base-100 px-5 py-2.5 flex items-center justify-between">
-        <button
-          onClick={() => setExpanded((v) => !v)}
-          className="flex items-center gap-1.5 text-xs font-medium text-base-content/70 hover:text-base-content transition-colors cursor-pointer"
-        >
-          <svg
-            className={`w-3.5 h-3.5 transition-transform duration-200 ${expanded ? "rotate-180" : ""}`}
-            fill="none"
-            viewBox="0 0 24 24"
-            stroke="currentColor"
-          >
-            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 9l-7 7-7-7" />
-          </svg>
-          {expanded ? "Hide tickets" : `View ${qty} ticket${qty !== 1 ? "s" : ""}`}
-        </button>
-        <Link
-          to={
-            qty > 1
-              ? `/tickets/${firstTicket.ticketCode}?codes=${tickets
-                  .map((t) => t.ticketCode)
-                  .filter(Boolean)
-                  .join(",")}`
-              : `/tickets/${firstTicket.ticketCode}`
-          }
-          className="text-xs font-medium text-base-content/70 hover:text-base-content flex items-center gap-1 transition-colors"
-        >
-          View Order
-          <svg className="w-3.5 h-3.5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
-          </svg>
-        </Link>
-      </div>
-
-      {/* Expanded ticket list */}
-      {expanded && (
-        <div className="px-5 pb-4 space-y-2">
-          {tickets.map((ticket) => (
-            <Link
-              key={ticket._id}
-              to={`/tickets/${ticket.ticketCode}`}
-              className="flex items-center gap-3 bg-base-100 rounded-xl px-4 py-3 hover:bg-base-200/60 transition-colors group"
-            >
-              <div className="w-8 h-8 rounded-lg bg-gradient-to-br from-primary to-primary/70 flex items-center justify-center flex-shrink-0">
-                <svg
-                  className="w-4 h-4 text-white"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={1.5}
-                    d="M15 5v2m0 4v2m0 4v2M5 5a2 2 0 00-2 2v3a2 2 0 110 4v3a2 2 0 002 2h14a2 2 0 002-2v-3a2 2 0 110-4V7a2 2 0 00-2-2H5z"
-                  />
-                </svg>
-              </div>
-              <div className="flex-1 min-w-0">
-                <p className="text-xs font-medium text-base-content font-mono">
-                  {ticket.ticketCode}
-                </p>
-                <p className="text-[10px] text-base-content/50">
-                  {formatCurrency(event?.ticketPrice ?? 0)}
-                  {ticket.checkedIn && " · Checked in"}
-                </p>
-              </div>
-              {ticket.checkedIn && (
-                <span className="text-xs font-medium px-2 py-0.5 rounded-full bg-blue-50 text-blue-700 border border-blue-200">
-                  ✓ Checked In
-                </span>
-              )}
-              <svg
-                className="w-3.5 h-3.5 text-base-content/30 group-hover:text-base-content/50 transition-colors flex-shrink-0"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M9 5l7 7-7 7"
-                />
-              </svg>
-            </Link>
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
-
-const CATEGORY_COLORS = {
-  Language: "from-blue-500 to-primary/70",
-  Religious: "from-emerald-500 to-teal-600",
-  Academic: "from-primary to-primary/70",
-  Arts: "from-primary to-primary/70",
-  Other: "from-amber-500 to-orange-600",
-};
-
-function EnrollmentRow({ enrollment }) {
-  const [expanded, setExpanded] = useState(false);
-  const [cancelling, setCancelling] = useState(false);
-  const [cancelDone, setCancelDone] = useState(enrollment.subscriptionStatus === "cancelled");
-  const [reactivating, setReactivating] = useState(false);
-  const [participants, setParticipants] = useState(enrollment.participants || []);
-  const [removingIdx, setRemovingIdx] = useState(null);
-  const [editingId, setEditingId] = useState(null);
-  const [editForm, setEditForm] = useState({ name: "", age: "" });
-  const [savingEdit, setSavingEdit] = useState(false);
-  const [confirm, setConfirm] = useState(null);
-  const [toast, setToast] = useState(null);
-  const [editingPhone, setEditingPhone] = useState(false);
-  const [phoneValue, setPhoneValue] = useState(enrollment.buyerPhone || "");
-  const [savingPhone, setSavingPhone] = useState(false);
-  const [phoneError, setPhoneError] = useState("");
-  const [currentPhone, setCurrentPhone] = useState(enrollment.buyerPhone || "");
-  const course = enrollment.courseId;
-  if (!course) return null;
-  const gradient = CATEGORY_COLORS[course.category] || CATEGORY_COLORS.Other;
-  const hasParticipants = participants.length > 0;
-  const isSubscription = !!enrollment.subscriptionId;
-  const periodEnd = enrollment.currentPeriodEnd ? new Date(enrollment.currentPeriodEnd) : null;
-
-  const showToast = (message, type = "error") => {
-    setToast({ message, type });
-    setTimeout(() => setToast(null), 4000);
-  };
-
-  const handleSavePhone = async () => {
-    setPhoneError("");
-    if (!phoneValue.trim()) {
-      setPhoneError("Phone number is required.");
-      return;
-    }
-    if (!validatePhone(phoneValue)) {
-      setPhoneError("Please enter a valid UK phone number.");
-      return;
-    }
-    setSavingPhone(true);
-    try {
-      const res = await fetchWithAuth(
-        `${import.meta.env.VITE_DEV_URI}courses/enrollments/${enrollment._id}`,
-        {
-          method: "PUT",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({ buyerPhone: phoneValue.trim() }),
-        }
-      );
-      const data = await res.json();
-      if (!res.ok) throw new Error(data.error || "Failed to update phone");
-      setCurrentPhone(phoneValue.trim());
-      setEditingPhone(false);
-      showToast("Phone number updated", "success");
-    } catch (err) {
-      setPhoneError(err.message);
-    } finally {
-      setSavingPhone(false);
-    }
-  };
-
-  const handleCancel = () => {
-    setConfirm({
-      title: "Cancel subscription",
-      message: periodEnd
-        ? `Are you sure you want to cancel? You'll keep access until ${periodEnd.toLocaleDateString("en-GB", { day: "numeric", month: "short", year: "numeric" })}.`
-        : "Are you sure you want to cancel? You'll keep access until the end of your current billing period.",
-      confirmText: "Yes, cancel",
-      variant: "danger",
-      onConfirm: async () => {
-        setConfirm(null);
-        setCancelling(true);
-        try {
-          const res = await fetchWithAuth(
-            `${import.meta.env.VITE_DEV_URI}courses/enrollments/${enrollment._id}/cancel`,
-            { method: "POST" }
-          );
-          const data = await res.json();
-          if (res.ok) setCancelDone(true);
-          else showToast(data.error || "Failed to cancel");
-        } catch {
-          showToast("Something went wrong");
-        }
-        setCancelling(false);
-      },
-    });
-  };
-
-  const handleReactivate = async () => {
-    setReactivating(true);
-    try {
-      const res = await fetchWithAuth(
-        `${import.meta.env.VITE_DEV_URI}courses/enrollments/${enrollment._id}/reactivate`,
-        { method: "POST" }
-      );
-      const data = await res.json();
-      if (res.ok) {
-        if (data.url) {
-          window.location.href = data.url;
-          return;
-        }
-        setCancelDone(false);
-      } else {
-        showToast(data.error || "Failed to reactivate");
-      }
-    } catch {
-      showToast("Something went wrong");
-    }
-    setReactivating(false);
-  };
-
-  const handleRemoveParticipant = (index) => {
-    const name = participants[index]?.name || "this participant";
-    setConfirm({
-      title: "Remove participant",
-      message: `Remove ${name} from this course? This cannot be undone.`,
-      confirmText: "Remove",
-      variant: "danger",
-      onConfirm: async () => {
-        setConfirm(null);
-        setRemovingIdx(index);
-        try {
-          const res = await fetchWithAuth(
-            `${import.meta.env.VITE_DEV_URI}courses/enrollments/${enrollment._id}/remove-participant`,
-            {
-              method: "POST",
-              headers: { "Content-Type": "application/json" },
-              body: JSON.stringify({ participantId: participants[index]._id }),
-            }
-          );
-          const data = await res.json();
-          if (res.ok) {
-            setParticipants(data.participants);
-          } else {
-            showToast(data.error || "Failed to remove participant");
-          }
-        } catch {
-          showToast("Something went wrong");
-        }
-        setRemovingIdx(null);
-      },
-    });
-  };
-
-  const startEdit = (p) => {
-    setEditingId(p._id);
-    setEditForm({ name: p.name || "", age: p.age || "" });
-  };
-
-  const cancelEdit = () => {
-    setEditingId(null);
-    setEditForm({ name: "", age: "" });
-  };
-
-  const handleSaveEdit = async (participantId) => {
-    if (!editForm.name.trim()) {
-      showToast("Name cannot be empty");
-      return;
-    }
-    setSavingEdit(true);
-    try {
-      const res = await fetchWithAuth(
-        `${import.meta.env.VITE_DEV_URI}courses/enrollments/${enrollment._id}/participants/${participantId}`,
-        {
-          method: "PUT",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({
-            name: editForm.name.trim(),
-            age: editForm.age ? Number(editForm.age) : undefined,
-          }),
-        }
-      );
-      const data = await res.json();
-      if (res.ok) {
-        setParticipants(data.participants);
-        setEditingId(null);
-        showToast("Participant updated", "success");
-      } else {
-        showToast(data.error || "Failed to update participant");
-      }
-    } catch {
-      showToast("Something went wrong");
-    }
-    setSavingEdit(false);
-  };
-
-  return (
-    <div className="bg-white rounded-2xl border border-base-300 shadow-sm hover:shadow-md transition-all duration-200 overflow-hidden">
-      <div className="flex items-center">
-        {course.images?.[0] ? (
-          <img
-            src={optimizeCloudinaryUrl(course.images[0])}
-            alt={course.title}
-            className="w-28 flex-shrink-0 object-cover"
-            width="112"
-            height="88"
-            style={{ minHeight: 88 }}
-          />
-        ) : (
-          <div
-            className={`w-16 flex-shrink-0 bg-gradient-to-b ${gradient} flex items-center justify-center py-6`}
-          >
-            <svg
-              className="w-6 h-6 text-white/70"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={1.5}
-                d="M12 6.253v13m0-13C10.832 5.477 9.246 5 7.5 5S4.168 5.477 3 6.253v13C4.168 18.477 5.754 18 7.5 18s3.332.477 4.5 1.253m0-13C13.168 5.477 14.754 5 16.5 5c1.747 0 3.332.477 4.5 1.253v13C19.832 18.477 18.247 18 16.5 18c-1.746 0-3.332.477-4.5 1.253"
-              />
-            </svg>
-          </div>
-        )}
-        <div className="flex-1 px-5 py-4 flex flex-col justify-center min-w-0">
-          <p className="font-semibold text-base-content truncate text-base">{course.title}</p>
-          <p className="text-sm text-base-content/50 mt-0.5">
-            {course.instructor && <span>{course.instructor}</span>}
-            {course.city && <span> · {course.city}</span>}
-          </p>
-          {course.schedule && (
-            <p className="text-xs text-base-content/50 mt-1">{course.schedule}</p>
-          )}
-          <p className="text-xs text-base-content/50 mt-1 font-mono">
-            {enrollment.enrollmentCode ?? enrollment._id.slice(-8).toUpperCase()}
-            {" · "}
-            {participants.length} participant{participants.length !== 1 ? "s" : ""}
-            {isSubscription && ` · ${INTERVAL_ADJ[course.billingInterval] || "Monthly"}`}
-            {currentPhone && ` · ${currentPhone}`}
-          </p>
-        </div>
-        <div className="flex flex-col items-end justify-center px-5 gap-2 flex-shrink-0">
-          <span className="text-sm font-semibold text-base-content">
-            {course.price > 0 ? formatCurrency(course.price * participants.length) : "Free"}
-          </span>
-          {course.price > 0 && participants.length > 1 && (
-            <span className="text-[10px] text-base-content/50">
-              {formatCurrency(course.price)} x {participants.length}
-            </span>
-          )}
-          <span
-            className={`text-xs font-medium px-2.5 py-0.5 rounded-full border ${
-              enrollment.status === "active"
-                ? "bg-green-50 text-green-700 border-green-200"
-                : enrollment.status === "paid"
-                  ? "bg-green-50 text-green-700 border-green-200"
-                  : enrollment.status === "free"
-                    ? "bg-blue-50 text-blue-600 border-blue-200"
-                    : enrollment.status === "cancelled"
-                      ? "bg-orange-50 text-orange-600 border-orange-200"
-                      : enrollment.status === "past_due"
-                        ? "bg-red-50 text-red-600 border-red-200"
-                        : "bg-base-100 text-base-content/50 border-base-300"
-            }`}
-          >
-            {enrollment.status === "active"
-              ? "✓ Subscribed"
-              : enrollment.status === "paid"
-                ? "✓ Enrolled"
-                : enrollment.status === "free"
-                  ? "✓ Free"
-                  : enrollment.status === "cancelled"
-                    ? "Cancelled"
-                    : enrollment.status === "past_due"
-                      ? "⚠ Payment due"
-                      : enrollment.status}
-          </span>
-        </div>
-      </div>
-
-      {/* Enrollment info bar */}
-      {isSubscription && (
-        <div
-          className={`px-5 py-2.5 text-xs flex items-center justify-between border-t ${cancelDone ? "bg-orange-50 border-orange-100" : "bg-blue-50 border-blue-100"}`}
-        >
-          <div>
-            {cancelDone ? (
-              <span className="text-orange-600 font-medium">
-                ⏳ Cancelled — access until{" "}
-                {periodEnd
-                  ? periodEnd.toLocaleDateString("en-GB", {
-                      day: "numeric",
-                      month: "short",
-                      year: "numeric",
-                    })
-                  : "end of period"}
-              </span>
-            ) : (
-              <span className="text-blue-600">
-                🔄 {INTERVAL_ADJ[course.billingInterval] || "Monthly"} subscription
-                {periodEnd && (
-                  <span className="text-blue-400 ml-1">
-                    · renews{" "}
-                    {periodEnd.toLocaleDateString("en-GB", { day: "numeric", month: "short" })}
-                  </span>
-                )}
-              </span>
-            )}
-          </div>
-          {cancelDone ? (
-            <button
-              onClick={handleReactivate}
-              disabled={reactivating}
-              className="text-xs text-green-600 hover:text-green-800 font-medium hover:underline transition-colors disabled:opacity-50 cursor-pointer disabled:cursor-not-allowed"
-            >
-              {reactivating ? "Reactivating..." : "Reactivate"}
-            </button>
-          ) : (
-            <button
-              onClick={handleCancel}
-              disabled={cancelling}
-              className="text-xs text-red-500 hover:text-red-700 font-medium hover:underline transition-colors disabled:opacity-50 cursor-pointer disabled:cursor-not-allowed"
-            >
-              {cancelling ? "Cancelling..." : "Cancel enrollment"}
-            </button>
-          )}
-        </div>
-      )}
-
-      {/* Phone edit */}
-      <div className="px-5 py-2.5 border-t border-base-100">
-        {editingPhone ? (
-          <div className="space-y-2">
-            <label className="text-xs font-medium text-base-content/70">Phone</label>
-            <input
-              type="tel"
-              value={phoneValue}
-              onChange={(e) => setPhoneValue(e.target.value)}
-              className="input input-sm input-bordered w-full text-xs"
-              placeholder="Phone (07...)"
-              autoComplete="tel"
-            />
-            {phoneError && <p className="text-red-500 text-xs">{phoneError}</p>}
-            <div className="flex gap-2">
-              <button
-                onClick={handleSavePhone}
-                disabled={savingPhone}
-                className="btn btn-xs btn-primary text-[10px]"
-              >
-                {savingPhone ? "Saving..." : "Save"}
-              </button>
-              <button
-                onClick={() => {
-                  setEditingPhone(false);
-                  setPhoneValue(currentPhone);
-                  setPhoneError("");
-                }}
-                className="btn btn-xs btn-ghost text-[10px]"
-              >
-                Cancel
-              </button>
-            </div>
-          </div>
-        ) : (
-          <div className="flex items-center gap-2">
-            <span className="text-xs text-base-content/50">Phone: {currentPhone || "Not set"}</span>
-            {!cancelDone && (
-              <button
-                onClick={() => setEditingPhone(true)}
-                className="text-xs text-primary hover:text-primary/80 font-medium cursor-pointer"
-              >
-                {currentPhone ? "Edit" : "Add"}
-              </button>
-            )}
-          </div>
-        )}
-      </div>
-
-      {/* Footer — participants toggle + view course */}
-      <div className="flex items-center justify-between border-t border-base-100 px-5 py-2.5">
-        {hasParticipants ? (
-          <button
-            onClick={() => setExpanded((v) => !v)}
-            className="flex items-center gap-1.5 text-xs font-medium text-base-content/70 hover:text-base-content transition-colors"
-          >
-            <svg
-              className={`w-3.5 h-3.5 transition-transform ${expanded ? "rotate-180" : ""}`}
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M19 9l-7 7-7-7"
-              />
-            </svg>
-            {expanded
-              ? "Hide participants"
-              : `${participants.length} participant${participants.length !== 1 ? "s" : ""}`}
-          </button>
-        ) : (
-          <span className="text-xs text-base-content/50">No participants recorded</span>
-        )}
-        <Link
-          to={`/courses/${toSlug(course.title, course._id)}`}
-          aria-label={`View ${course.title} course`}
-          className="text-xs font-medium text-base-content/70 hover:text-base-content flex items-center gap-1 transition-colors"
-        >
-          View Course
-          <svg className="w-3.5 h-3.5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
-          </svg>
-        </Link>
-      </div>
-
-      {/* Participants list */}
-      {expanded && hasParticipants && (
-        <div className="px-5 pb-4 grid grid-cols-1 sm:grid-cols-2 gap-2">
-          {participants.map((p, i) =>
-            editingId === p._id ? (
-              <div
-                key={p._id || i}
-                className="flex flex-col gap-1.5 bg-base-100 rounded-xl px-3 py-2"
-              >
-                <input
-                  type="text"
-                  value={editForm.name}
-                  onChange={(e) => setEditForm({ ...editForm, name: e.target.value })}
-                  placeholder="Name"
-                  className="input input-xs input-bordered w-full text-xs"
-                />
-                <input
-                  type="number"
-                  value={editForm.age}
-                  onChange={(e) => setEditForm({ ...editForm, age: e.target.value })}
-                  placeholder="Age"
-                  className="input input-xs input-bordered w-20 text-xs"
-                />
-                <div className="flex gap-1.5 justify-end">
-                  <button
-                    onClick={cancelEdit}
-                    disabled={savingEdit}
-                    className="btn btn-xs btn-ghost text-[10px]"
-                  >
-                    Cancel
-                  </button>
-                  <button
-                    onClick={() => handleSaveEdit(p._id)}
-                    disabled={savingEdit}
-                    className="btn btn-xs btn-primary text-[10px]"
-                  >
-                    {savingEdit ? "Saving..." : "Save"}
-                  </button>
-                </div>
-              </div>
-            ) : (
-              <div
-                key={p._id || i}
-                className="flex items-center gap-2.5 bg-base-100 rounded-xl px-3 py-2 group/participant"
-              >
-                <div className="w-6 h-6 rounded-full bg-gradient-to-br from-primary to-primary/70 text-white text-[10px] font-bold flex items-center justify-center flex-shrink-0">
-                  {p.name?.[0]?.toUpperCase() ?? "?"}
-                </div>
-                <div
-                  className="min-w-0 flex-1 cursor-pointer"
-                  onClick={() => !cancelDone && startEdit(p)}
-                  title="Click to edit"
-                >
-                  <p className="text-xs font-medium text-base-content truncate">{p.name}</p>
-                  <div className="flex gap-2 text-[10px] text-base-content/50">
-                    {p.age && <span>Age {p.age}</span>}
-                    {p.email && <span className="truncate">{p.email}</span>}
-                  </div>
-                </div>
-                {!cancelDone && (
-                  <button
-                    onClick={() => startEdit(p)}
-                    title={`Edit ${p.name}`}
-                    className="opacity-0 group-hover/participant:opacity-100 transition-opacity text-base-content/40 hover:text-primary flex-shrink-0 cursor-pointer"
-                  >
-                    <svg
-                      className="w-3.5 h-3.5"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      stroke="currentColor"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M15.232 5.232l3.536 3.536m-2.036-5.036a2.5 2.5 0 113.536 3.536L6.5 21.036H3v-3.572L16.732 3.732z"
-                      />
-                    </svg>
-                  </button>
-                )}
-                {participants.length > 1 && !cancelDone && (
-                  <button
-                    onClick={() => handleRemoveParticipant(i)}
-                    disabled={removingIdx !== null}
-                    title={`Remove ${p.name}`}
-                    className="opacity-0 group-hover/participant:opacity-100 transition-opacity text-red-400 hover:text-red-600 disabled:opacity-30 flex-shrink-0 cursor-pointer disabled:cursor-not-allowed"
-                  >
-                    {removingIdx === i ? (
-                      <div className="w-4 h-4 rounded-full border-2 border-red-300 border-t-red-500 animate-spin" />
-                    ) : (
-                      <svg
-                        className="w-4 h-4"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                        stroke="currentColor"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          strokeWidth={2}
-                          d="M6 18L18 6M6 6l12 12"
-                        />
-                      </svg>
-                    )}
-                  </button>
-                )}
-              </div>
-            )
-          )}
-        </div>
-      )}
-
-      {/* Confirm modal */}
-      {confirm && (
-        <div
-          className="fixed inset-0 z-50 flex items-center justify-center bg-black/40 backdrop-blur-sm"
-          onClick={() => setConfirm(null)}
-        >
-          <div
-            className="bg-white rounded-2xl shadow-xl border border-base-300 max-w-sm w-full mx-4 p-6 animate-[scaleIn_0.15s_ease-out]"
-            onClick={(e) => e.stopPropagation()}
-          >
-            <div className="flex items-center gap-3 mb-3">
-              <div
-                className={`w-10 h-10 rounded-full flex items-center justify-center flex-shrink-0 ${
-                  confirm.variant === "danger" ? "bg-red-50" : "bg-base-200"
-                }`}
-              >
-                {confirm.variant === "danger" ? (
-                  <svg
-                    className="w-5 h-5 text-red-500"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={1.5}
-                      d="M12 9v3.75m-9.303 3.376c-.866 1.5.217 3.374 1.948 3.374h14.71c1.73 0 2.813-1.874 1.948-3.374L13.949 3.378c-.866-1.5-3.032-1.5-3.898 0L2.697 16.126zM12 15.75h.007v.008H12v-.008z"
-                    />
-                  </svg>
-                ) : (
-                  <svg
-                    className="w-5 h-5 text-base-content/70"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={1.5}
-                      d="M9.879 7.519c1.171-1.025 3.071-1.025 4.242 0 1.172 1.025 1.172 2.687 0 3.712-.203.179-.43.326-.67.442-.745.361-1.45.999-1.45 1.827v.75M21 12a9 9 0 11-18 0 9 9 0 0118 0zm-9 5.25h.008v.008H12v-.008z"
-                    />
-                  </svg>
-                )}
-              </div>
-              <h3 className="text-base font-semibold text-base-content">{confirm.title}</h3>
-            </div>
-            <p className="text-sm text-base-content/50 mb-6 ml-[52px]">{confirm.message}</p>
-            <div className="flex gap-3 justify-end">
-              <button
-                onClick={() => setConfirm(null)}
-                className="px-4 py-2 text-sm font-medium text-base-content/70 bg-base-200 hover:bg-base-300 rounded-xl transition-colors cursor-pointer"
-              >
-                Cancel
-              </button>
-              <button
-                onClick={confirm.onConfirm}
-                className={`px-4 py-2 text-sm font-medium text-white rounded-xl transition-colors cursor-pointer ${
-                  confirm.variant === "danger"
-                    ? "bg-red-500 hover:bg-red-600"
-                    : "bg-primary hover:bg-primary/80"
-                }`}
-              >
-                {confirm.confirmText || "Confirm"}
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Toast notification */}
-      {toast && (
-        <div className="fixed bottom-6 left-1/2 -translate-x-1/2 z-50 animate-[slideUp_0.2s_ease-out]">
-          <div
-            className={`flex items-center gap-2.5 px-5 py-3 rounded-xl shadow-lg border text-sm font-medium ${
-              toast.type === "error"
-                ? "bg-red-50 text-red-700 border-red-200"
-                : "bg-green-50 text-green-700 border-green-200"
-            }`}
-          >
-            {toast.type === "error" ? (
-              <svg
-                className="w-4 h-4 flex-shrink-0"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M6 18L18 6M6 6l12 12"
-                />
-              </svg>
-            ) : (
-              <svg
-                className="w-4 h-4 flex-shrink-0"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M5 13l4 4L19 7"
-                />
-              </svg>
-            )}
-            {toast.message}
-            <button
-              onClick={() => setToast(null)}
-              className="ml-2 text-current opacity-50 hover:opacity-100 cursor-pointer"
-            >
-              <svg className="w-3.5 h-3.5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M6 18L18 6M6 6l12 12"
-                />
-              </svg>
-            </button>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-const BOOKING_STATUS_STYLES = {
-  confirmed: "bg-green-50 text-green-700 border-green-200",
-  pending: "bg-yellow-50 text-yellow-700 border-yellow-200",
-  completed: "bg-blue-50 text-blue-600 border-blue-200",
-  cancelled: "bg-red-50 text-red-600 border-red-200",
-};
-
-function VenueBookingRow({ booking }) {
-  const [expanded, setExpanded] = useState(false);
-  const slot = booking.slot;
-  const venue = booking.venue;
-  const status = booking.status;
-
-  const slotDateStr = slot?.date
-    ? new Date(slot.date).toLocaleDateString("en-GB", {
-        weekday: "short",
-        day: "numeric",
-        month: "short",
-        year: "numeric",
-      })
-    : "—";
-
-  return (
-    <div className="bg-white rounded-2xl border border-base-300 shadow-sm hover:shadow-md transition-all duration-200 overflow-hidden">
-      <div className="flex">
-        {/* Date block */}
-        <div className="w-16 flex-shrink-0 bg-gradient-to-b from-primary to-primary/70 flex flex-col items-center justify-center text-white py-4">
-          <span className="text-xs font-semibold uppercase opacity-80">
-            {slot?.date ? new Date(slot.date).toLocaleString("en-GB", { month: "short" }) : "—"}
-          </span>
-          <span className="text-2xl font-bold leading-tight">
-            {slot?.date ? new Date(slot.date).getDate() : "—"}
-          </span>
-        </div>
-
-        {/* Info */}
-        <div className="flex-1 px-5 py-4 flex flex-col justify-center min-w-0">
-          <p className="font-semibold text-base-content truncate text-base">
-            {venue?.name ?? "Venue"}
-          </p>
-          <p className="text-sm text-base-content/50 mt-0.5">
-            {slotDateStr}
-            {slot?.startTime && slot?.endTime && (
-              <span>
-                {" "}
-                · {slot.startTime}–{slot.endTime}
-              </span>
-            )}
-            {venue?.city && <span> · {venue.city}</span>}
-          </p>
-          {booking.eventName && booking.eventName !== "N/A" && (
-            <p className="text-xs text-base-content/50 mt-1">{booking.eventName}</p>
-          )}
-          <p className="text-xs text-base-content/50 mt-1 font-mono">
-            {booking.bookingCode ?? `#${booking._id.slice(-8).toUpperCase()}`} ·{" "}
-            {booking.numberOfAttendees} attendee{booking.numberOfAttendees !== 1 ? "s" : ""}
-          </p>
-        </div>
-
-        {/* Price + status */}
-        <div className="flex flex-col items-end justify-center px-5 gap-2 flex-shrink-0">
-          <span className="text-sm font-semibold text-base-content">
-            {formatCurrency(booking.totalPrice)}
-          </span>
-          <span
-            className={`text-xs font-medium px-2.5 py-0.5 rounded-full border capitalize ${
-              BOOKING_STATUS_STYLES[status] ?? "bg-base-200 text-base-content/50 border-base-300"
-            }`}
-          >
-            {status}
-          </span>
-        </div>
-      </div>
-
-      {/* Expand toggle */}
-      <div className="border-t border-base-100 px-5 py-2.5">
-        <button
-          onClick={() => setExpanded((v) => !v)}
-          className="flex items-center gap-1.5 text-xs font-medium text-base-content/70 hover:text-base-content transition-colors cursor-pointer"
-        >
-          <svg
-            className={`w-3.5 h-3.5 transition-transform duration-200 ${expanded ? "rotate-180" : ""}`}
-            fill="none"
-            viewBox="0 0 24 24"
-            stroke="currentColor"
-          >
-            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 9l-7 7-7-7" />
-          </svg>
-          {expanded ? "Hide details" : "View details"}
-        </button>
-      </div>
-
-      {/* Expanded details */}
-      {expanded && (
-        <div className="px-5 pb-5 space-y-3">
-          <div className="grid grid-cols-2 gap-x-6 gap-y-2 text-sm">
-            <div>
-              <p className="text-xs text-base-content/40 uppercase tracking-wide font-medium mb-0.5">
-                Venue
-              </p>
-              <p className="text-base-content font-medium">{venue?.name ?? "—"}</p>
-              {(venue?.street || venue?.city) && (
-                <p className="text-xs text-base-content/50">
-                  {[venue.street, venue.city].filter(Boolean).join(", ")}
-                </p>
-              )}
-            </div>
-            <div>
-              <p className="text-xs text-base-content/40 uppercase tracking-wide font-medium mb-0.5">
-                Date &amp; Time
-              </p>
-              <p className="text-base-content">{slotDateStr}</p>
-              {slot?.startTime && slot?.endTime && (
-                <p className="text-xs text-base-content/50">
-                  {slot.startTime} – {slot.endTime}
-                </p>
-              )}
-            </div>
-            <div>
-              <p className="text-xs text-base-content/40 uppercase tracking-wide font-medium mb-0.5">
-                Attendees
-              </p>
-              <p className="text-base-content">{booking.numberOfAttendees}</p>
-            </div>
-            <div>
-              <p className="text-xs text-base-content/40 uppercase tracking-wide font-medium mb-0.5">
-                Total Paid
-              </p>
-              <p className="text-base-content font-medium">{formatCurrency(booking.totalPrice)}</p>
-            </div>
-            {booking.eventName && booking.eventName !== "N/A" && (
-              <div>
-                <p className="text-xs text-base-content/40 uppercase tracking-wide font-medium mb-0.5">
-                  Event Name
-                </p>
-                <p className="text-base-content">{booking.eventName}</p>
-              </div>
-            )}
-            {booking.eventDescription && (
-              <div className="col-span-2">
-                <p className="text-xs text-base-content/40 uppercase tracking-wide font-medium mb-0.5">
-                  Description
-                </p>
-                <p className="text-base-content/70 text-xs">{booking.eventDescription}</p>
-              </div>
-            )}
-            <div>
-              <p className="text-xs text-base-content/40 uppercase tracking-wide font-medium mb-0.5">
-                Status
-              </p>
-              <span
-                className={`text-xs font-medium px-2 py-0.5 rounded-full border capitalize ${
-                  BOOKING_STATUS_STYLES[status] ??
-                  "bg-base-200 text-base-content/50 border-base-300"
-                }`}
-              >
-                {status}
-              </span>
-            </div>
-            <div>
-              <p className="text-xs text-base-content/40 uppercase tracking-wide font-medium mb-0.5">
-                Reference
-              </p>
-              <p className="font-mono text-xs text-base-content/70">
-                {booking.bookingCode ?? `#${booking._id.slice(-8).toUpperCase()}`}
-              </p>
-            </div>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function EventSubscriptionRow({ subscription, onAction }) {
-  const [cancelling, setCancelling] = useState(false);
-  const [cancelDone, setCancelDone] = useState(subscription.subscriptionStatus === "cancelled");
-  const [reactivating, setReactivating] = useState(false);
-  const [confirm, setConfirm] = useState(null);
-  const [toast, setToast] = useState(null);
-  const event = subscription.eventId;
-  if (!event) return null;
-
-  const periodEnd = subscription.currentPeriodEnd ? new Date(subscription.currentPeriodEnd) : null;
-  const interval = event.subscriptionInterval || "month";
-
-  const showToast = (message, type = "error") => {
-    setToast({ message, type });
-    setTimeout(() => setToast(null), 4000);
-  };
-
-  const handleCancel = () => {
-    setConfirm({
-      title: "Cancel subscription",
-      message: periodEnd
-        ? `Are you sure you want to cancel? You'll keep access until ${periodEnd.toLocaleDateString("en-GB", { day: "numeric", month: "short", year: "numeric" })}.`
-        : "Are you sure you want to cancel? You'll keep access until the end of your current billing period.",
-      confirmText: "Yes, cancel",
-      variant: "danger",
-      onConfirm: async () => {
-        setConfirm(null);
-        setCancelling(true);
-        try {
-          const res = await fetchWithAuth(
-            `${import.meta.env.VITE_DEV_URI}events/subscriptions/${subscription._id}/cancel`,
-            { method: "POST" }
-          );
-          const data = await res.json();
-          if (res.ok) {
-            setCancelDone(true);
-            onAction();
-          } else showToast(data.error || "Failed to cancel");
-        } catch {
-          showToast("Something went wrong");
-        }
-        setCancelling(false);
-      },
-    });
-  };
-
-  const handleReactivate = async () => {
-    setReactivating(true);
-    try {
-      const res = await fetchWithAuth(
-        `${import.meta.env.VITE_DEV_URI}events/subscriptions/${subscription._id}/reactivate`,
-        { method: "POST" }
-      );
-      const data = await res.json();
-      if (res.ok) {
-        if (data.url) {
-          window.location.href = data.url;
-          return;
-        }
-        setCancelDone(false);
-        onAction();
-      } else {
-        showToast(data.error || "Failed to reactivate");
-      }
-    } catch {
-      showToast("Something went wrong");
-    }
-    setReactivating(false);
-  };
-
-  return (
-    <div className="bg-white rounded-2xl border border-base-300 shadow-sm hover:shadow-md transition-all duration-200 overflow-hidden">
-      <div className="flex items-center">
-        {event.images?.[0] ? (
-          <img
-            src={optimizeCloudinaryUrl(event.images[0])}
-            alt={event.title}
-            className="w-28 flex-shrink-0 object-cover"
-            width="112"
-            height="88"
-            style={{ minHeight: 88 }}
-          />
-        ) : (
-          <div className="w-16 flex-shrink-0 bg-gradient-to-b from-primary to-primary/70 flex items-center justify-center py-6">
-            <svg
-              className="w-6 h-6 text-white/70"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={1.5}
-                d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15"
-              />
-            </svg>
-          </div>
-        )}
-        <div className="flex-1 px-5 py-4 flex flex-col justify-center min-w-0">
-          <p className="font-semibold text-base-content truncate text-base">{event.title}</p>
-          <p className="text-sm text-base-content/50 mt-0.5">
-            {event.dayOfWeek && (
-              <span>{event.dayOfWeek.charAt(0).toUpperCase() + event.dayOfWeek.slice(1)}s</span>
-            )}
-            {event.openingTime && <span> · {event.openingTime}</span>}
-            {event.city && <span> · {event.city}</span>}
-          </p>
-          <p className="text-xs text-base-content/50 mt-1 font-mono">
-            {INTERVAL_ADJ[interval] || "Monthly"} · £{event.ticketPrice?.toFixed(2)}/{interval}
-          </p>
-        </div>
-        <div className="flex flex-col items-end justify-center px-5 gap-2 flex-shrink-0">
-          <span className="text-sm font-semibold text-base-content">
-            {formatCurrency(event.ticketPrice)}
-          </span>
-          <span
-            className={`text-xs font-medium px-2.5 py-0.5 rounded-full border ${
-              subscription.status === "active" && !cancelDone
-                ? "bg-green-50 text-green-700 border-green-200"
-                : subscription.status === "past_due"
-                  ? "bg-red-50 text-red-600 border-red-200"
-                  : "bg-orange-50 text-orange-600 border-orange-200"
-            }`}
-          >
-            {subscription.status === "active" && !cancelDone
-              ? "✓ Subscribed"
-              : subscription.status === "past_due"
-                ? "⚠ Payment due"
-                : "Cancelled"}
-          </span>
-        </div>
-      </div>
-
-      {/* Enrollment info bar */}
-      <div
-        className={`px-5 py-2.5 text-xs flex items-center justify-between border-t ${cancelDone ? "bg-orange-50 border-orange-100" : "bg-blue-50 border-blue-100"}`}
-      >
-        <div>
-          {cancelDone ? (
-            <span className="text-orange-600 font-medium">
-              Cancelled — access until{" "}
-              {periodEnd
-                ? periodEnd.toLocaleDateString("en-GB", {
-                    day: "numeric",
-                    month: "short",
-                    year: "numeric",
-                  })
-                : "end of period"}
-            </span>
-          ) : (
-            <span className="text-blue-600">
-              {INTERVAL_ADJ[interval] || "Monthly"} enrollment
-              {periodEnd && (
-                <span className="text-blue-400 ml-1">
-                  · renews{" "}
-                  {periodEnd.toLocaleDateString("en-GB", {
-                    day: "numeric",
-                    month: "short",
-                  })}
-                </span>
-              )}
-            </span>
-          )}
-        </div>
-        {cancelDone ? (
-          <button
-            onClick={handleReactivate}
-            disabled={reactivating}
-            className="text-xs text-green-600 hover:text-green-800 font-medium hover:underline transition-colors disabled:opacity-50 cursor-pointer disabled:cursor-not-allowed"
-          >
-            {reactivating ? "Reactivating..." : "Reactivate"}
-          </button>
-        ) : (
-          <button
-            onClick={handleCancel}
-            disabled={cancelling}
-            className="text-xs text-red-500 hover:text-red-700 font-medium hover:underline transition-colors disabled:opacity-50 cursor-pointer disabled:cursor-not-allowed"
-          >
-            {cancelling ? "Cancelling..." : "Cancel subscription"}
-          </button>
-        )}
-      </div>
-
-      {/* Footer — view event */}
-      <div className="flex items-center justify-end border-t border-base-100 px-5 py-2.5">
-        <Link
-          to={`/events/${toSlug(event.title, event._id)}`}
-          aria-label={`View ${event.title} event`}
-          className="text-xs font-medium text-base-content/70 hover:text-base-content flex items-center gap-1 transition-colors"
-        >
-          View Event
-          <svg className="w-3.5 h-3.5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
-          </svg>
-        </Link>
-      </div>
-
-      {/* Confirm modal */}
-      {confirm && (
-        <div
-          className="fixed inset-0 z-50 flex items-center justify-center bg-black/40 backdrop-blur-sm"
-          onClick={() => setConfirm(null)}
-        >
-          <div
-            className="bg-white rounded-2xl shadow-xl border border-base-300 max-w-sm w-full mx-4 p-6 animate-[scaleIn_0.15s_ease-out]"
-            onClick={(e) => e.stopPropagation()}
-          >
-            <div className="flex items-center gap-3 mb-3">
-              <div className="w-10 h-10 rounded-full flex items-center justify-center flex-shrink-0 bg-red-50">
-                <svg
-                  className="w-5 h-5 text-red-500"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={1.5}
-                    d="M12 9v3.75m-9.303 3.376c-.866 1.5.217 3.374 1.948 3.374h14.71c1.73 0 2.813-1.874 1.948-3.374L13.949 3.378c-.866-1.5-3.032-1.5-3.898 0L2.697 16.126zM12 15.75h.007v.008H12v-.008z"
-                  />
-                </svg>
-              </div>
-              <h3 className="text-base font-semibold text-base-content">{confirm.title}</h3>
-            </div>
-            <p className="text-sm text-base-content/50 mb-6 ml-[52px]">{confirm.message}</p>
-            <div className="flex gap-3 justify-end">
-              <button
-                onClick={() => setConfirm(null)}
-                className="px-4 py-2 text-sm font-medium text-base-content/70 bg-base-200 hover:bg-base-300 rounded-xl transition-colors cursor-pointer"
-              >
-                Cancel
-              </button>
-              <button
-                onClick={confirm.onConfirm}
-                className="px-4 py-2 text-sm font-medium text-white rounded-xl transition-colors cursor-pointer bg-red-500 hover:bg-red-600"
-              >
-                {confirm.confirmText || "Confirm"}
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Toast */}
-      {toast && (
-        <div className="fixed bottom-6 left-1/2 -translate-x-1/2 z-50 animate-[slideUp_0.2s_ease-out]">
-          <div
-            className={`flex items-center gap-2.5 px-5 py-3 rounded-xl shadow-lg border text-sm font-medium ${
-              toast.type === "error"
-                ? "bg-red-50 text-red-700 border-red-200"
-                : "bg-green-50 text-green-700 border-green-200"
-            }`}
-          >
-            {toast.message}
-            <button
-              onClick={() => setToast(null)}
-              className="ml-2 text-current opacity-50 hover:opacity-100 cursor-pointer"
-            >
-              <svg className="w-3.5 h-3.5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M6 18L18 6M6 6l12 12"
-                />
-              </svg>
-            </button>
-          </div>
-        </div>
-      )}
     </div>
   );
 }
