+++ src/pages/content/Home.jsx
@@ -3,6 +3,8 @@ import { Helmet } from "react-helmet-async";
 import { Link } from "react-router-dom";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import ConfirmModal from "../../components/common/ConfirmModal";
+import ImageWithFallback from "../../components/common/ImageWithFallback";
+import PageEditBar from "../../components/common/PageEditBar";
 import CourseCard from "../../components/courses/CourseCard";
 import EventCard from "../../components/events/EventCard";
 import Button from "../../components/ui/Button";
@@ -37,7 +39,7 @@ function SectionHeader({ kicker, title, action }) {
   );
 }
 
-export default function Home() {
+export default function Home({ previewContent = null }) {
   const { data: events } = useEvents();
   const { data: courses } = useCourses();
   const queryClient = useQueryClient();
@@ -48,10 +50,12 @@ export default function Home() {
   const [saving, setSaving] = useState(false);
   const [resetting, setResetting] = useState(false);
   const [saveError, setSaveError] = useState(null);
+  const [submitSuccess, setSubmitSuccess] = useState(null);
   const [confirmOpen, setConfirmOpen] = useState(false);
   const heroImageInputRef = useRef(null);
 
-  const canEdit = isAdmin() || isModerator();
+  const canEdit = !previewContent && (isAdmin() || isModerator());
+  const canEditDirectly = isAdmin();
 
   const { data: rawPageContent } = useQuery({
     queryKey: queryKeys.pageContent.home,
@@ -59,10 +63,12 @@ export default function Home() {
       const r = await fetch(`${API}pageContent/home`);
       return r.json();
     },
+    enabled: !previewContent,
   });
 
-  const pageContent =
-    rawPageContent && Object.keys(rawPageContent).length > 0
+  const pageContent = previewContent
+    ? mergeWithDefaults(previewContent)
+    : rawPageContent && Object.keys(rawPageContent).length > 0
       ? mergeWithDefaults(rawPageContent)
       : DEFAULTS;
 
@@ -71,6 +77,7 @@ export default function Home() {
     setHeroImagePreview(null);
     setHeroImageFile(null);
     setSaveError(null);
+    setSubmitSuccess(null);
     setEditing(true);
   };
 
@@ -106,6 +113,7 @@ export default function Home() {
 
   const handleSave = async () => {
     setSaveError(null);
+    setSubmitSuccess(null);
 
     const errors = [];
     if (!draft.heroTitle?.trim()) errors.push("Hero title cannot be empty.");
@@ -129,8 +137,9 @@ export default function Home() {
       );
       if (heroImageFile) formData.append("heroImage", heroImageFile);
 
-      const res = await fetchWithAuth(`${API}pageContent/home`, {
-        method: "PUT",
+      const url = canEditDirectly ? `${API}pageContent/home` : `${API}pageContentRequests/home`;
+      const res = await fetchWithAuth(url, {
+        method: canEditDirectly ? "PUT" : "POST",
         body: formData,
       });
       if (!res.ok) {
@@ -138,13 +147,23 @@ export default function Home() {
         throw new Error(body?.message || `Server error (${res.status})`);
       }
 
-      const { pageContent: saved } = await res.json();
-      queryClient.setQueryData(queryKeys.pageContent.home, saved);
-      setEditing(false);
+      if (canEditDirectly) {
+        const { pageContent: saved } = await res.json();
+        queryClient.setQueryData(queryKeys.pageContent.home, saved);
+        setEditing(false);
+      } else {
+        setSubmitSuccess("Your change request has been submitted for admin approval.");
+        setEditing(false);
+      }
       setHeroImagePreview(null);
       setHeroImageFile(null);
     } catch (err) {
-      setSaveError(err.message || "Failed to save changes. Please try again.");
+      setSaveError(
+        err.message ||
+          (canEditDirectly
+            ? "Failed to save changes. Please try again."
+            : "Failed to submit request. Please try again.")
+      );
     } finally {
       setSaving(false);
     }
@@ -173,7 +192,7 @@ export default function Home() {
       </Helmet>
 
       {canEdit && (
-        <div className="sticky top-20 z-40 mx-auto mb-6 mt-6 flex w-[min(1200px,calc(100%-2rem))] flex-wrap justify-end gap-2 rounded-3xl border border-white/60 bg-white/75 px-4 py-3 shadow-[var(--shadow-soft)] backdrop-blur-xl">
+        <PageEditBar>
           {!editing ? (
             <Button variant="primary" onClick={handleEdit}>
               Edit Page
@@ -183,26 +202,34 @@ export default function Home() {
               <Button variant="secondary" onClick={handleCancel}>
                 Cancel
               </Button>
-              <button
-                onClick={() => setConfirmOpen(true)}
-                disabled={resetting}
-                className="btn border border-warning/25 bg-warning/10 text-warning-content hover:bg-warning/15"
-              >
-                {resetting ? "Resetting..." : "Reset to Defaults"}
-              </button>
+              {canEditDirectly && (
+                <button
+                  onClick={() => setConfirmOpen(true)}
+                  disabled={resetting}
+                  className="btn border border-warning/25 bg-warning/10 text-warning-content hover:bg-warning/15"
+                >
+                  {resetting ? "Resetting..." : "Reset to Defaults"}
+                </button>
+              )}
               <Button variant="primary" onClick={handleSave} disabled={saving}>
-                {saving ? "Saving..." : "Save Changes"}
+                {saving
+                  ? canEditDirectly
+                    ? "Saving..."
+                    : "Submitting..."
+                  : canEditDirectly
+                    ? "Save Changes"
+                    : "Submit for Approval"}
               </Button>
             </>
           )}
-        </div>
+        </PageEditBar>
       )}
 
       {saveError && (
         <div className="fixed right-3 top-28 z-50 max-w-[calc(100vw-1.5rem)] rounded-2xl border border-red-200 bg-red-50/95 px-4 py-3 text-sm text-red-600 shadow-lg backdrop-blur-sm animate-scale-in sm:right-6 sm:top-36 sm:max-w-sm">
           <div className="flex items-start gap-2">
             <svg
-              className="mt-0.5 h-4 w-4 flex-shrink-0"
+              className="mt-0.5 h-4 w-4 shrink-0"
               fill="none"
               viewBox="0 0 24 24"
               stroke="currentColor"
@@ -220,7 +247,7 @@ export default function Home() {
             </div>
             <button
               onClick={() => setSaveError(null)}
-              className="ml-auto flex-shrink-0 text-red-400 hover:text-red-600"
+              className="ml-auto shrink-0 text-red-400 hover:text-red-600"
             >
               <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                 <path
@@ -235,10 +262,47 @@ export default function Home() {
         </div>
       )}
 
-      <section className="page-section py-6 md:py-10">
-        <div className="grid items-center gap-8 lg:grid-cols-[1.15fr_0.85fr]">
+      {submitSuccess && (
+        <div className="fixed right-3 top-28 z-50 max-w-[calc(100vw-1.5rem)] rounded-2xl border border-green-200 bg-green-50/95 px-4 py-3 text-sm text-green-700 shadow-lg backdrop-blur-sm animate-scale-in sm:right-6 sm:top-36 sm:max-w-sm">
+          <div className="flex items-start gap-2">
+            <svg
+              className="mt-0.5 h-4 w-4 shrink-0"
+              fill="none"
+              viewBox="0 0 24 24"
+              stroke="currentColor"
+            >
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                strokeWidth={2}
+                d="M5 13l4 4L19 7"
+              />
+            </svg>
+            <div>
+              <p className="mb-0.5 font-medium">Request submitted</p>
+              <p className="text-xs leading-relaxed">{submitSuccess}</p>
+            </div>
+            <button
+              onClick={() => setSubmitSuccess(null)}
+              className="ml-auto shrink-0 text-green-400 hover:text-green-600"
+            >
+              <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth={2}
+                  d="M6 18L18 6M6 6l12 12"
+                />
+              </svg>
+            </button>
+          </div>
+        </div>
+      )}
+
+      <section className="page-section ">
+        <div className="grid items-stretch gap-2 mt-6 min-h-128 lg:grid-cols-[1.15fr_0.85fr]">
           <div
-            className={`hero-panel rounded-[2rem] p-6 sm:p-8 lg:p-10 ${
+            className={`hero-panel flex flex-col justify-center gap-4 rounded-4xl p-6 sm:p-8 lg:p-10 ${
               editing ? "ring-2 ring-secondary/35" : ""
             }`}
           >
@@ -248,14 +312,14 @@ export default function Home() {
             {editing ? (
               <>
                 <input
-                  className="glass-input mb-4 text-3xl font-bold text-primary"
+                  className="glass-input  text-3xl font-bold text-primary"
                   value={draft.heroTitle}
                   onChange={(e) => setDraft({ ...draft, heroTitle: e.target.value })}
                   placeholder="Hero title"
                 />
                 <textarea
                   rows={5}
-                  className="glass-input mb-4 resize-none text-base"
+                  className="glass-input  resize-none text-base"
                   value={draft.heroDescription}
                   onChange={(e) => setDraft({ ...draft, heroDescription: e.target.value })}
                   placeholder="Hero description"
@@ -287,18 +351,18 @@ export default function Home() {
             </div>
           </div>
 
-          <div className="relative">
+          <div className="relative flex items-center">
             <div className="absolute -left-4 -top-4 h-24 w-24 rounded-full bg-secondary/12 blur-2xl" />
             <div className="absolute -bottom-6 -right-3 h-28 w-28 rounded-full bg-primary/15 blur-2xl" />
             <div
-              className={`glass-card relative overflow-hidden rounded-[2rem] p-3 ${
+              className={`w-full relative overflow-hidden rounded-4xl p-3 ${
                 editing ? "ring-2 ring-primary/20" : ""
               }`}
             >
-              <img
+              <ImageWithFallback
                 src={heroImage}
                 alt="Community event"
-                className="aspect-[4/4.2] w-full rounded-[1.5rem] object-cover sm:aspect-[4/3.8]"
+                className="aspect-[4/4.2] w-full rounded-3xl object-cover sm:aspect-[4/3.8]"
                 fetchpriority="high"
                 width="500"
                 height="375"
@@ -334,94 +398,98 @@ export default function Home() {
         </div>
       </section>
 
-      <section className="page-section py-4 md:py-8">
-        <div className="grid gap-4 rounded-[2rem] bg-neutral px-6 py-8 text-white shadow-[var(--shadow-strong)] sm:grid-cols-2 lg:grid-cols-4">
-          {[
-            { value: "500+", label: "Community members engaged" },
-            { value: "50+", label: "Events delivered" },
-            { value: "20+", label: "Courses offered" },
-            { value: "Leeds", label: "Serving the local community" },
-          ].map(({ value, label }) => (
-            <div
-              key={label}
-              className="rounded-2xl border border-white/8 bg-white/5 p-4 text-center"
-            >
-              <div className="mb-2 text-3xl font-semibold tracking-[-0.04em] text-white">
-                {value}
-              </div>
-              <div className="text-sm text-neutral-content/70">{label}</div>
+      {!previewContent && (
+        <>
+          <section className="page-section py-4 md:py-8">
+            <div className="grid gap-4 rounded-4xl bg-neutral px-6 py-8 text-white shadow-(--shadow-strong) sm:grid-cols-2 lg:grid-cols-4">
+              {[
+                { value: "500+", label: "Community members engaged" },
+                { value: "50+", label: "Events delivered" },
+                { value: "20+", label: "Courses offered" },
+                { value: "Leeds", label: "Serving the local community" },
+              ].map(({ value, label }) => (
+                <div
+                  key={label}
+                  className="rounded-4xl border border-white/8 bg-white/5 p-4 text-center"
+                >
+                  <div className="mb-2 text-3xl font-semibold tracking-[-0.04em] text-white">
+                    {value}
+                  </div>
+                  <div className="text-sm text-neutral-content/70">{label}</div>
+                </div>
+              ))}
             </div>
-          ))}
-        </div>
-      </section>
+          </section>
 
-      <section className="page-section py-8 md:py-12">
-        <SectionHeader
-          kicker="Community Calendar"
-          title="Upcoming Events"
-          action={
-            <Link
-              to="/events/asc"
-              className="self-center md:self-auto inline-flex items-center gap-2 rounded-full border border-base-300 bg-white/80 px-5 py-3 text-sm font-semibold text-base-content shadow-sm transition-all hover:-translate-y-0.5 hover:border-primary/20 hover:text-primary"
-            >
-              View All Events
-              <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M13 7l5 5m0 0l-5 5m5-5H6"
-                />
-              </svg>
-            </Link>
-          }
-        />
-        {upcomingEvents.length === 0 ? (
-          <div className="glass-card rounded-[1.75rem] p-8 text-center">
-            <p className="text-base-content/70">No upcoming events found.</p>
-          </div>
-        ) : (
-          <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
-            {upcomingEvents.map((event) => (
-              <EventCard key={event._id} event={event} />
-            ))}
-          </div>
-        )}
-      </section>
+          <section className="page-section py-8 md:py-12">
+            <SectionHeader
+              kicker="Community Calendar"
+              title="Upcoming Events"
+              action={
+                <Link
+                  to="/events/asc"
+                  className="self-center md:self-auto inline-flex items-center gap-2 rounded-full border border-base-300 bg-white/80 px-5 py-3 text-sm font-semibold text-base-content shadow-sm transition-all hover:-translate-y-0.5 hover:border-primary/20 hover:text-primary"
+                >
+                  View All Events
+                  <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth={2}
+                      d="M13 7l5 5m0 0l-5 5m5-5H6"
+                    />
+                  </svg>
+                </Link>
+              }
+            />
+            {upcomingEvents.length === 0 ? (
+              <div className="glass-card rounded-[1.75rem] p-8 text-center">
+                <p className="text-base-content/70">No upcoming events found.</p>
+              </div>
+            ) : (
+              <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
+                {upcomingEvents.map((event) => (
+                  <EventCard key={event._id} event={event} />
+                ))}
+              </div>
+            )}
+          </section>
 
-      <section className="page-section py-8 md:py-12">
-        <SectionHeader
-          kicker="Learning Opportunities"
-          title="Available Courses"
-          action={
-            <Link
-              to="/courses"
-              className="self-center md:self-auto inline-flex items-center gap-2 rounded-full border border-base-300 bg-white/80 px-5 py-3 text-sm font-semibold text-base-content shadow-sm transition-all hover:-translate-y-0.5 hover:border-primary/20 hover:text-primary"
-            >
-              View All Courses
-              <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M13 7l5 5m0 0l-5 5m5-5H6"
-                />
-              </svg>
-            </Link>
-          }
-        />
-        {availableCourses.length === 0 ? (
-          <div className="glass-card rounded-[1.75rem] p-8 text-center">
-            <p className="text-base-content/70">No courses available right now.</p>
-          </div>
-        ) : (
-          <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
-            {availableCourses.map((course) => (
-              <CourseCard key={course._id} course={course} />
-            ))}
-          </div>
-        )}
-      </section>
+          <section className="page-section py-8 md:py-12">
+            <SectionHeader
+              kicker="Learning Opportunities"
+              title="Available Courses"
+              action={
+                <Link
+                  to="/courses"
+                  className="self-center md:self-auto inline-flex items-center gap-2 rounded-full border border-base-300 bg-white/80 px-5 py-3 text-sm font-semibold text-base-content shadow-sm transition-all hover:-translate-y-0.5 hover:border-primary/20 hover:text-primary"
+                >
+                  View All Courses
+                  <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth={2}
+                      d="M13 7l5 5m0 0l-5 5m5-5H6"
+                    />
+                  </svg>
+                </Link>
+              }
+            />
+            {availableCourses.length === 0 ? (
+              <div className="glass-card rounded-[1.75rem] p-8 text-center">
+                <p className="text-base-content/70">No courses available right now.</p>
+              </div>
+            ) : (
+              <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
+                {availableCourses.map((course) => (
+                  <CourseCard key={course._id} course={course} />
+                ))}
+              </div>
+            )}
+          </section>
+        </>
+      )}
     </div>
   );
 }
