+++ src/util/pageContentRequestStatus.js
@@ -1,0 +1,5 @@
+export const REQUEST_STATUS_STYLES = {
+  pending: "bg-yellow-50 text-yellow-700 border-yellow-200",
+  approved: "bg-green-50 text-green-700 border-green-200",
+  declined: "bg-red-50 text-red-600 border-red-200",
+};
